clean sample 03
