+++ 03-Session-Handling/app/src/main/AndroidManifest.xml
@@ -13,14 +13,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity
-            android:name=".activities.StartActivity"
-            android:noHistory="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
 
         <!-- Auth0 Lock -->
         <activity
@@ -37,17 +29,23 @@
 
                 <data
                     android:host="@string/auth0_domain"
-                    android:pathPrefix="/android/YOUR APP PATH/callback"
+                    android:pathPrefix="/android/com.auth0.sessiondemo/callback"
                     android:scheme="https" />
             </intent-filter>
         </activity>
-        <!-- Auth0 Lock End -->
 
-        <activity android:name=".activities.LockActivity" />
-        <activity android:name=".activities.MainActivity" />
         <activity
             android:name="com.auth0.android.provider.WebAuthActivity"
             android:theme="@style/Lock.Theme" />
+        <!-- Auth0 Lock End -->
+
+        <activity android:name=".LoginActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity android:name=".MainActivity" />
     </application>
 
 </manifest>
