+++ 03-Session-Handling/app/src/main/java/com/auth0/sessiondemo/utils/Constants.java
@@ -1,10 +1,8 @@
 package com.auth0.sessiondemo.utils;
 
-public class Constants {
-
-    public final static String REFRESH_TOKEN = "AUTH0_REFRESH_TOKEN";
-    public final static String ACCESS_TOKEN = "AUTH0_ACCESS_TOKEN";
-    public final static String ID_TOKEN = "AUTH0_ID_TOKEN";
-    public final static String CREDENTIAL_TYPE = "AUTH0_CREDENTIAL_TYPE";
-
+class Constants {
+    final static String REFRESH_TOKEN = "AUTH0_REFRESH_TOKEN";
+    final static String ACCESS_TOKEN = "AUTH0_ACCESS_TOKEN";
+    final static String ID_TOKEN = "AUTH0_ID_TOKEN";
+    final static String CREDENTIAL_TYPE = "AUTH0_CREDENTIAL_TYPE";
 }
