--- 03-Session-Handling/app/src/main/res/layout/activity_start.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="@dimen/activity_general_margin"
-    android:gravity="center"
-    tools:context=".activities.StartActivity">
-
-</RelativeLayout>
