+++ 03-Session-Handling/app/src/main/res/values/strings.xml
@@ -1,14 +1,10 @@
 <resources>
     <string name="app_name">03 - Session Management</string>
-
-    <string name="auth0_client_id">{CLIENT_ID}</string>
-    <string name="auth0_domain">{DOMAIN}</string>
-
+    <string name="auth0_preferences">AUTH0_SHARED_PREFERENCES_KEY</string>
     <string name="refresh">Get New ID With a RefreshTokenID</string>
     <string name="token">Get New ID With a TokenID</string>
     <string name="logout">Logout</string>
 
-    <string name="auth0_preferences">AUTH0_SHARED_PREFERENCES_KEY</string>
-
-
+    <string name="auth0_client_id">{CLIENT_ID}</string>
+    <string name="auth0_domain">{DOMAIN}</string>
 </resources>
